moved entity extraction to own module
