+++ modules/nlu/src/backend/engine2/predict-pipeline.ts
@@ -4,9 +4,11 @@ import _ from 'lodash'
 import { getClosestToken } from '../pipelines/language/ft_featurizer'
 import LanguageIdentifierProvider, { NA_LANG } from '../pipelines/language/ft_lid'
 import * as math from '../tools/math'
+import { PatternEntity } from '../typings'
 
 import CRFExtractor2 from './crf-extractor2'
-import { extractUtteranceEntities, PatternEntity, Tools } from './engine2'
+import { Tools } from './engine2'
+import { extractUtteranceEntities } from './entity-extractor'
 import { Model } from './model-service'
 import { EXACT_MATCH_STR_OPTIONS, ExactMatchIndex, Intent, TrainArtefacts } from './training-pipeline'
 import Utterance, { buildUtterances } from './utterance'
