+++ modules/nlu/src/backend/typings.ts
@@ -162,4 +162,35 @@ export interface BotState {
   trainWatcher: sdk.ListenHandle
 }
 
-export type TFIDF = Dic<number>
+export type TFIDF = _.Dictionary<number>
+
+export type PatternEntity = Readonly<{
+  name: string
+  pattern: string
+  examples: string[]
+  ignoreCase: boolean
+  sensitive: boolean
+}>
+
+export type ListEntity = Readonly<{
+  name: string
+  synonyms: { [canonical: string]: string[] }
+  fuzzyMatching: boolean
+  sensitive: boolean
+}>
+
+export type ListEntityModel = Readonly<{
+  type: 'custom.list'
+  id: string
+  languageCode: string
+  entityName: string
+  fuzzyMatching: boolean
+  sensitive: boolean
+  /** @example { 'Air Canada': [ ['Air', '_Canada'], ['air', 'can'] ] } */
+  mappingsTokens: _.Dictionary<string[][]>
+}>
+
+// add value in extracted slots
+export type ExtractedSlot = { confidence: number; name: string; source: any }
+export type ExtractedEntity = { confidence: number; type: string; metadata: any; value: string }
+export type EntityExtractionResult = ExtractedEntity & { start: number; end: number }
